extract stdlib.lua parts to separate internal modules
